tanzania ecaas rice dataset fixed
