+++ .gitignore
@@ -84,6 +84,7 @@ celerybeat-schedule
 # virtualenv
 venv/
 ENV/
+harvest-env/
 
 # Spyder project settings
 .spyderproject
