+++ process_labels/loading_funcs/__init__.py
@@ -7,14 +7,13 @@ from .central_asia import load_central_asia
 from .rwanda import load_rwanda_ceo
 from .kenya import load_kenya, load_kenya_non_crop
 from .uganda import load_uganda
-from .tanzania import load_tanzania
+from .tanzania import load_tanzania, load_tanzania_ecaas
 from .croplands import load_croplands
 from .zimbabwe import load_zimbabwe
 from .mali import load_mali, load_mali_crop_noncrop
 from .france import load_ile_de_france, load_reunion, load_martinique
 from .canada import load_canada
 from .germany import load_germany
-from .tanzania_ecaas import load_tanzania_ecaas
 
 
 __all__ = [
